Move src/{api,->}/{code_asset,data_asset,asset}.dart, update changelog,...
