+++ pkgs/native_assets_cli/lib/src/api/test.dart
@@ -9,12 +9,12 @@ import 'package:test/test.dart';
 import 'package:yaml/yaml.dart';
 
 import '../architecture.dart';
+import '../asset.dart';
 import '../build_mode.dart';
 import '../c_compiler_config.dart';
 import '../ios_sdk.dart';
 import '../link_mode_preference.dart';
 import '../os.dart';
-import 'asset.dart';
 import 'build_config.dart';
 import 'build_output.dart';
