+++ pkgs/native_assets_cli/lib/src/data_asset.dart
