+++ pkgs/native_assets_cli/lib/src/link_mode.dart
@@ -2,8 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'api/asset.dart';
 import 'api/build_config.dart';
+import 'asset.dart';
 
 /// The link mode for a [CodeAsset].
 ///
